refactor : separate style
