+++ src/App.js
@@ -25,6 +25,8 @@ function App() {
 const AppWrapper = tw.div`
   flex
   relative
+  bg-black
+  min-h-screen
 `;
 
 export default App;
